extensions sync and rename add -> install
